refactor: refactor config loading and remove arrayable resolution
- Rename `src/util/config.ts` to `src/config.ts` and `src/util/config.test.ts` to `src/config.test.ts`
- Update import paths due to the file rename
- Remove the `resolveArrayable` function
- Rename `loadUserConfig` to `loadDefaultConfigFile` and create a new `loadUserConfig` function that loads either a specified config file or a default one
- Update debug messages to use `%O` for object formatting
