+++ packages/cross-release-cli/src/config.ts
@@ -3,8 +3,8 @@ import { isFileSync } from "@rainbowatcher/fs-extra"
 import { toAbsolute } from "@rainbowatcher/path-extra"
 import defu from "defu"
 import { loadConfig } from "unconfig"
-import createDebug from "../util/debug"
-import type { Arrayable, ReleaseOptions, ResolvedOptions } from "../types"
+import createDebug from "./util/debug"
+import type { ReleaseOptions, ResolvedOptions } from "./types"
 
 const debug = createDebug("config")
 
@@ -21,11 +21,6 @@ export function resolveAltOptions<K extends keyof ReleaseOptions>(
         : { ..._defaultValue, ...value as any }
 }
 
-export function resolveArrayable<T = any>(maybeArr: Arrayable<T>): T[] {
-    if (!maybeArr) return []
-    return Array.isArray(maybeArr) ? maybeArr : [maybeArr]
-}
-
 export async function loadUserSpecifiedConfigFile(configPath: string, currentOpts: Partial<ReleaseOptions>): Promise<Partial<ReleaseOptions>> {
     const absConfigPath = toAbsolute(configPath)
     if (!isFileSync(absConfigPath)) {
@@ -38,11 +33,11 @@ export async function loadUserSpecifiedConfigFile(configPath: string, currentOpt
         }],
     })
 
-    debug("load specified config file:", sources)
+    debug("load specified config file: %O", sources)
     return defu({ config: toAbsolute(currentOpts.config ?? "") }, currentOpts, config)
 }
 
-export async function loadUserConfig(cwd = process.cwd()): Promise<Partial<ReleaseOptions>> {
+export async function loadDefaultConfigFile(cwd = process.cwd()): Promise<Partial<ReleaseOptions>> {
     const { config: userConfig, sources } = await loadConfig<Partial<ReleaseOptions>>({
         cwd,
         sources: [
@@ -58,6 +53,17 @@ export async function loadUserConfig(cwd = process.cwd()): Promise<Partial<Relea
     })
 
     debug("load user config", sources)
-    debug("user config:", userConfig)
+    debug("user config: %O", userConfig)
+    return userConfig
+}
+
+
+export async function loadUserConfig(opts: ReleaseOptions) {
+    let userConfig: Partial<ReleaseOptions>
+    if (opts.config) {
+        userConfig = await loadUserSpecifiedConfigFile(opts.config, opts)
+    } else {
+        userConfig = await loadDefaultConfigFile(opts.cwd)
+    }
     return userConfig
 }
